fixed error after delet trash.json

+++ smart_rm/remove.py
@@ -7,6 +7,7 @@ import datetime
 import os
 import tarfile
 import json
+import sys
 
 
 class RemoveToTrash:
@@ -14,17 +15,38 @@ class RemoveToTrash:
         self.data = []
         self.trash_dir = trash_dir
 
+    def _reformatting_json(self):
+        logging.debug('start _reformatting_json')
+        files = []
+        for i in os.listdir(self.trash_dir):
+            if i != 'trash.json':
+                files.append(i)
+        remove_index = []
+        for index, file in enumerate(self.data):
+            if not file['name_archive'] in files:
+                remove_index.append(index)
+        for index in remove_index:
+            del self.data[index]
+        logging.debug('_reformatting_json completed')
+
     def _load_data(self):
         logging.debug('start _load_data')
         if not os.path.isdir(self.trash_dir):
             os.mkdir(self.trash_dir)
-
+        if os.path.isdir(self.trash_dir):
+            files = []
+            for i in os.listdir(self.trash_dir):
+                files.append(i)
+            if not os.path.isfile(f'{self.trash_dir}/trash.json') and files != []:
+                sys.exit()
         for i in os.listdir(self.trash_dir):
             if i == 'trash.json':
                 try:
                     with open(self.trash_dir + '/trash.json', 'r') as file:
                         self.data = json.load(file)
+                        self._reformatting_json()
                         break
+
                 except:
                     logging.error('нет доступа к trash.json')
         else:
@@ -34,11 +56,12 @@ class RemoveToTrash:
     def _save_data(self):
         logging.debug('start _save_data')
         try:
+            self._reformatting_json()
             with open(self.trash_dir + '/trash.json', 'w') as file:
                 json.dump(self.data, file)
+            logging.debug('_save_data completed')
         except:
             logging.error('нет доступа к trash.json')
-        logging.debug('_save_data completed')
 
     def _file_address(self, link):
         folder = []
@@ -69,29 +92,32 @@ class RemoveToTrash:
         name = os.path.basename(link)
         name_archive = f'{name}_{time}.tar.gz'
         link_archive = f'{self.trash_dir}/{name_archive}'
-        with tarfile.open(f'{self.trash_dir}/{name_archive}', 'w:gz') as archive:
-            archive.add(name)
-        if os.path.isdir(link):
-            file, directories = self._file_address(link)
-            for i in file:
-                os.remove(i)
-                logging.debug(f'remove file : {i}')
-            for i in directories:
-                os.rmdir(i)
-                logging.debug(f'remove folder : {i}')
-            os.rmdir(link)
-        else:
-            os.remove(link)
-            logging.debug(f'remove file : {link}')
+        if os.path.isdir(link) or os.path.isfile(link):
+            with tarfile.open(f'{self.trash_dir}/{name_archive}', 'w:gz') as archive:
+                archive.add(name)
+            if os.path.isdir(link):
+                file, directories = self._file_address(link)
+                for i in file:
+                    os.remove(i)
+                    logging.debug(f'remove file : {i}')
+                for i in directories:
+                    os.rmdir(i)
+                    logging.debug(f'remove folder : {i}')
+                os.rmdir(link)
+            else:
+                os.remove(link)
+                logging.debug(f'remove file : {link}')
 
-        size = os.path.getsize(link_archive)
-        self.data.append({'name': name, 'size': size, 'link_to_file': link_to_file,
-                          'time_stamp': time_stamp, 'link': link,
-                          'name_archive': name_archive,
-                          'link_archive': link_archive,
-                          'id_trash': id_trash})
-        self._save_data()
-        logging.debug('remove completed')
+            size = os.path.getsize(link_archive)
+            self.data.append({'name': name, 'size': size, 'link_to_file': link_to_file,
+                              'time_stamp': time_stamp, 'link': link,
+                              'name_archive': name_archive,
+                              'link_archive': link_archive,
+                              'id_trash': id_trash})
+            self._save_data()
+            logging.debug('remove completed')
+        else:
+            return 'error 3'
 
     def recovery(self, id_trash):
         logging.debug('start recovery')
